requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
idna==2.10
pyhpeimc==1.0.49
PyYAML==5.3.1
requests==2.25.0
typer==0.3.2
urllib3==1.26.2
tabulate==0.8.7
